+++ tests/test_AppManager.py
@@ -32,6 +32,11 @@ class Test_App:
         with pytest.raises(NotImplementedError):
             an_app.pages()
 
+    def test_default_tasks_method_returns_empty_list(self):
+
+        an_app = App()
+        assert an_app.tasks() == []
+
 
 class Test_AppManager_apps:
 
@@ -119,7 +124,7 @@ class Test_AppManager_apps:
         an_app_manager.add_app(an_app, make_current=True)
         assert an_app_manager.current_app is an_app
 
-    def test_when_app_made_curent_then_current_app_updatred(self, an_app_manager, an_app):
+    def test_when_app_made_curent_then_current_app_updated(self, an_app_manager, an_app):
 
         an_app_manager.add_app(an_app, make_current=False)
         an_app_manager.set_current_app(an_app)
@@ -132,40 +137,63 @@ class Test_AppManager_apps:
         with pytest.raises(ValueError):
             an_app_manager.set_current_app(an_app)
 
-    def test_when_app_added_without_make_current_then_wm_pages_unchanged(
+    def test_when_app_added_without_make_current_then_wm_pages_and_os_tasks_unchanged(
         self, a_wm, an_app_factory
     ):
 
         expected_pages = a_wm.pages()
+        expected_tasks = a_wm.os.tasks()
         expected_current = a_wm.current_page
         am = AppManager(a_wm)
         am.add_app(an_app_factory())
         assert a_wm.pages() == expected_pages
+        assert a_wm.os.tasks() == expected_tasks
         assert a_wm.current_page == expected_current
 
-    def test_when_app_added_with_make_current_true_then_wm_pages_updated_and_first_made_current(
+    def test_when_app_added_with_make_current_true_then_wm_pages_updated_and_first_made_current_and_tasks_added(
         self, a_wm, an_app_factory
     ):
 
+        existing_tasks = a_wm.os.tasks()
         an_app = an_app_factory()
         app_pages = an_app.pages()
+        app_tasks = an_app.tasks()
         am = AppManager(a_wm)
         am.add_app(an_app, make_current=True)
+        new_tasks = a_wm.os.tasks()[len(existing_tasks) :]
+        assert len(new_tasks) == (len(app_tasks) + len(app_pages))
+        # There should be the first two, the second two are the page tasks
+        os_app_tasks = new_tasks[0 : len(app_tasks)]
+        assert [t.fn for t in os_app_tasks] == an_app.t_fn
+        assert [t.touch_forces_execution for t in os_app_tasks] == an_app.t_tfe
+        assert [t.execution_interval_us for t in os_app_tasks] == an_app.t_ens
         assert a_wm.pages() == app_pages
         assert a_wm.current_page == app_pages[0]
 
-    def test_when_app_made_current_then_wm_pages_updated_and_first_made_current(
+    def test_when_app_made_current_then_wm_pages_and_tasks_updated_and_first_made_current(
         self, a_wm, an_app_factory
     ):
 
+        existing_tasks = a_wm.os.tasks()
         app_a = an_app_factory()
+        app_a_tasks = app_a.tasks()
         app_b = an_app_factory()
+        app_b_tasks = app_b.tasks()
         am = AppManager(a_wm)
         am.add_app(app_a, make_current=True)
         am.add_app(app_b)
+        # Check tasks are for app_a
+        new_tasks = a_wm.os.tasks()[len(existing_tasks) :]
+        os_app_tasks = new_tasks[0 : len(app_a_tasks)]
+        assert [t.fn for t in os_app_tasks] == app_a.t_fn
         assert a_wm.pages() == app_a.pages()
         assert a_wm.current_page == app_a.pages()[0]
         am.set_current_app(app_b)
+        # Check app_a task removed and app_b tasks present
+        assert not any(t.fn in app_a.t_fn for t in a_wm.os.tasks())
+        new_tasks = a_wm.os.tasks()[len(existing_tasks) :]
+        os_app_tasks = new_tasks[0 : len(app_b_tasks)]
+        assert [t.fn for t in os_app_tasks] == app_b.t_fn
         assert a_wm.pages() == app_b.pages()
         assert a_wm.current_page == app_b.pages()[0]
 
@@ -190,7 +218,7 @@ def a_wm():
 @pytest.fixture
 def an_app_factory():
     """
-    A factory to make mock apps with 2 pages
+    A factory to make mock apps with 2 pages and two tasks.
     """
 
     def make():
@@ -202,6 +230,14 @@ def an_app_factory():
         app.name = "Some app"
         app.description = "A desc"
         app.pages.return_value = (p1, p2)
+        app.t_fn = [mock.Mock(), mock.Mock()]
+        app.t_hz = [1, None]
+        app.t_ens = [1e6, -1]
+        app.t_tfe = [True, False]
+        app.tasks.return_value = (
+            App.Task(app.t_fn[0], app.t_hz[0], app.t_tfe[0]),
+            App.Task(app.t_fn[1], app.t_hz[1], app.t_tfe[1]),
+        )
         return app
 
     return make